refactor(infrastructure): inline data mappers, constants, and caching logic
