+++ src/modules/diet/food/domain/apiFoodRepository.ts
@@ -1,4 +1,4 @@
-import { type ApiFood } from '~/modules/diet/food/infrastructure/api/domain/apiFoodSchema'
+import { type ApiFood } from '~/modules/diet/food/domain/apiFood'
 
 export type ApiFoodRepository = {
   fetchApiFoods: () => Promise<readonly ApiFood[]>
