+++ src/modules/diet/food/infrastructure/api/infrastructure/supabase/supabaseFoodRepository.ts
@@ -1,14 +1,73 @@
-import { type Food, type NewFood } from '~/modules/diet/food/domain/food'
+import {
+  type Food,
+  foodSchema,
+  type NewFood,
+} from '~/modules/diet/food/domain/food'
 import {
   type FoodRepository,
   type FoodSearchParams,
 } from '~/modules/diet/food/domain/foodRepository'
-import { supabaseFoodMapper } from '~/modules/diet/food/infrastructure/api/infrastructure/supabase/supabaseFoodMapper'
-import { SUPABASE_TABLE_FOODS } from '~/modules/diet/food/infrastructure/supabase/constants'
+import { createMacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
+import { type Database, type Json } from '~/shared/supabase/database.types'
 import { supabase } from '~/shared/supabase/supabase'
 import { isSupabaseDuplicateEanError } from '~/shared/supabase/supabaseErrorUtils'
 import { wrapErrorWithStack } from '~/shared/utils/errorUtils'
 import { logging } from '~/shared/utils/logging'
+import { parseWithStack } from '~/shared/utils/parseWithStack'
+
+const SUPABASE_TABLE_FOODS = 'foods'
+
+type FoodDTO = Database['public']['Tables']['foods']['Row']
+type InsertFoodDTO = Database['public']['Tables']['foods']['Insert']
+
+function macrosToDomain(macrosDTO: Json) {
+  if (
+    macrosDTO === null ||
+    !(typeof macrosDTO === 'object') ||
+    !('carbs' in macrosDTO) ||
+    !('protein' in macrosDTO) ||
+    !('fat' in macrosDTO) ||
+    typeof macrosDTO.carbs !== 'number' ||
+    typeof macrosDTO.protein !== 'number' ||
+    typeof macrosDTO.fat !== 'number'
+  ) {
+    throw new Error(
+      'macrosDTO is missing macros field: ' + JSON.stringify(macrosDTO),
+    )
+  }
+
+  return createMacroNutrients({
+    carbsInMg: macrosDTO.carbs * 1000,
+    proteinInMg: macrosDTO.protein * 1000,
+    fatInMg: macrosDTO.fat * 1000,
+  })
+}
+
+function macrosToDTO(macroNutrients: Food['macros']): Json {
+  return {
+    carbs: macroNutrients.carbsInMg / 1000,
+    protein: macroNutrients.proteinInMg / 1000,
+    fat: macroNutrients.fatInMg / 1000,
+  }
+}
+
+function foodToDomain(dto: FoodDTO): Food {
+  return parseWithStack(foodSchema, {
+    ...dto,
+    ean: dto.ean ?? null,
+    source: dto.source ?? undefined,
+    macros: macrosToDomain(dto.macros),
+  })
+}
+
+function foodToInsertDTO(newFood: NewFood): InsertFoodDTO {
+  return {
+    name: newFood.name,
+    ean: newFood.ean ?? null,
+    macros: macrosToDTO(newFood.macros),
+    source: newFood.source ?? null,
+  }
+}
 
 export function createSupabaseFoodRepository(): FoodRepository {
   return {
@@ -86,7 +145,7 @@ async function fetchFoodByEan(
  * @throws Error if not created or on API/validation error
  */
 async function insertFood(newFood: NewFood): Promise<Food> {
-  const insertDTO = supabaseFoodMapper.toInsertDTO(newFood)
+  const insertDTO = foodToInsertDTO(newFood)
   const { data, error } = await supabase
     .from(SUPABASE_TABLE_FOODS)
     .insert(insertDTO)
@@ -101,7 +160,7 @@ async function insertFood(newFood: NewFood): Promise<Food> {
     throw wrapErrorWithStack(error)
   }
 
-  return supabaseFoodMapper.toDomain(data)
+  return foodToDomain(data)
 }
 
 /**
@@ -112,7 +171,7 @@ async function insertFood(newFood: NewFood): Promise<Food> {
  * @throws Error if not created or on API/validation error
  */
 async function upsertFood(newFood: NewFood): Promise<Food> {
-  const createDTO = supabaseFoodMapper.toInsertDTO(newFood)
+  const createDTO = foodToInsertDTO(newFood)
   const { data: food, error } = await supabase
     .from(SUPABASE_TABLE_FOODS)
     .upsert(createDTO)
@@ -127,7 +186,7 @@ async function upsertFood(newFood: NewFood): Promise<Food> {
     throw wrapErrorWithStack(error)
   }
 
-  return supabaseFoodMapper.toDomain(food)
+  return foodToDomain(food)
 }
 
 async function fetchFoodsByName(
@@ -165,7 +224,7 @@ async function fetchFoodsByName(
         : 'enhanced search'
 
     logging.debug(`Found ${resultsCount} foods using ${searchType}`)
-    return result.data.map(supabaseFoodMapper.toDomain)
+    return result.data.map(foodToDomain)
   } catch (err) {
     logging.error('Food search error:', err)
     throw err
@@ -242,7 +301,7 @@ async function internalCachedSearchFoods(
   }
 
   logging.debug(`Found ${foods.length} foods`)
-  return foods.map(supabaseFoodMapper.toDomain)
+  return foods.map(foodToDomain)
 }
 
 /**
@@ -262,5 +321,5 @@ async function fetchFoodsByIds(ids: Food['id'][]): Promise<readonly Food[]> {
     throw wrapErrorWithStack(error)
   }
 
-  return foods.map(supabaseFoodMapper.toDomain)
+  return foods.map(foodToDomain)
 }
