--- src/modules/diet/food/infrastructure/supabase/constants.ts
@@ -1,1 +1,0 @@
-export const SUPABASE_TABLE_FOODS = 'foods'
