--- src/modules/diet/macro-profile/infrastructure/supabase/constants.ts
@@ -1,1 +1,0 @@
-export const SUPABASE_TABLE_MACRO_PROFILES = 'macro_profiles'
