+++ src/modules/diet/recent-food/infrastructure/recentFoodRepository.ts
@@ -3,13 +3,32 @@ import {
   type RecentFood,
 } from '~/modules/diet/recent-food/domain/recentFood'
 import { type RecentFoodRepository } from '~/modules/diet/recent-food/domain/recentFoodRepository'
-import { type RecentFoodGateway } from '~/modules/diet/recent-food/infrastructure/supabase/supabaseRecentFoodGateway'
 import { type Template } from '~/modules/diet/template/domain/template'
 import { type User } from '~/modules/user/domain/user'
 import { logging } from '~/shared/utils/logging'
 
+type RecentFoodGatewayLike = {
+  fetchByUserTypeAndReferenceId: (
+    userId: User['uuid'],
+    type: RecentFood['type'],
+    referenceId: number,
+  ) => Promise<RecentFood | null>
+  fetchUserRecentFoodsAsTemplates: (
+    userId: User['uuid'],
+    search: string,
+    opts?: { limit?: number },
+  ) => Promise<readonly Template[]>
+  insert: (input: NewRecentFood) => Promise<RecentFood | null>
+  update: (id: number, input: NewRecentFood) => Promise<RecentFood | null>
+  deleteByReference: (
+    userId: User['uuid'],
+    type: RecentFood['type'],
+    referenceId: number,
+  ) => Promise<boolean>
+}
+
 export function createRecentFoodRepository(
-  gateway: RecentFoodGateway,
+  gateway: RecentFoodGatewayLike,
 ): RecentFoodRepository {
   return {
     async fetchByUserTypeAndReferenceId(
