--- src/modules/diet/recent-food/infrastructure/supabase/constants.ts
@@ -1,1 +1,0 @@
-export const SUPABASE_TABLE_RECENT_FOODS = 'recent_foods'
