+++ src/modules/diet/recipe/infrastructure/recipeRepository.ts
@@ -1,14 +1,43 @@
+import { createSignal } from 'solid-js'
+
 import {
   type NewRecipe,
   type Recipe,
 } from '~/modules/diet/recipe/domain/recipe'
 import { type RecipeRepository } from '~/modules/diet/recipe/domain/recipeRepository'
-import { recipeCacheStore } from '~/modules/diet/recipe/infrastructure/signals/recipeCacheStore'
 import { createSupabaseRecipeGateway } from '~/modules/diet/recipe/infrastructure/supabase/supabaseRecipeGateway'
 import { type User } from '~/modules/user/domain/user'
 import { logging } from '~/shared/utils/logging'
 
 const supabaseGateway = createSupabaseRecipeGateway()
+const [recipes, setRecipes] = createSignal<readonly Recipe[]>([])
+
+function upsertToCache(recipe: Recipe) {
+  setRecipes((current) => {
+    const existingIndex = current.findIndex((currentRecipe) => {
+      return currentRecipe.id === recipe.id
+    })
+
+    if (existingIndex >= 0) {
+      const updated = [...current]
+      updated[existingIndex] = recipe
+      return updated
+    }
+
+    return [...current, recipe]
+  })
+}
+
+function removeFromCache(criteria: { by: 'id'; value: Recipe['id'] }) {
+  setRecipes((current) =>
+    current.filter((recipe) => recipe.id !== criteria.value),
+  )
+}
+
+function findInCache(criteria: { by: 'id'; value: Recipe['id'] }) {
+  const current = recipes()
+  return current.find((recipe) => recipe.id === criteria.value) ?? null
+}
 
 export function createRecipeRepository(): RecipeRepository {
   return {
@@ -27,7 +56,7 @@ async function fetchUserRecipes(
   try {
     const recipes = await supabaseGateway.fetchUserRecipes(userId)
     for (const recipe of recipes) {
-      recipeCacheStore.upsertToCache(recipe)
+      upsertToCache(recipe)
     }
     return recipes
   } catch (error) {
@@ -39,21 +68,21 @@ async function fetchUserRecipes(
 async function fetchRecipeById(recipeId: Recipe['id']): Promise<Recipe | null> {
   try {
     // Check cache first
-    const cached = recipeCacheStore.findInCache({ by: 'id', value: recipeId })
+    const cached = findInCache({ by: 'id', value: recipeId })
     if (cached !== null) {
       return cached
     }
 
     const recipe = await supabaseGateway.fetchRecipeById(recipeId)
     if (recipe === null) {
-      recipeCacheStore.removeFromCache({ by: 'id', value: recipeId })
+      removeFromCache({ by: 'id', value: recipeId })
       return null
     }
-    recipeCacheStore.upsertToCache(recipe)
+    upsertToCache(recipe)
     return recipe
   } catch (error) {
     logging.error('Recipe error:', error)
-    recipeCacheStore.removeFromCache({ by: 'id', value: recipeId })
+    removeFromCache({ by: 'id', value: recipeId })
     return null
   }
 }
@@ -65,7 +94,7 @@ async function fetchUserRecipeByName(
   try {
     const recipes = await supabaseGateway.fetchUserRecipeByName(userId, name)
     for (const recipe of recipes) {
-      recipeCacheStore.upsertToCache(recipe)
+      upsertToCache(recipe)
     }
     return recipes
   } catch (error) {
@@ -78,7 +107,7 @@ async function insertRecipe(newRecipe: NewRecipe): Promise<Recipe | null> {
   try {
     const insertedRecipe = await supabaseGateway.insertRecipe(newRecipe)
     if (insertedRecipe !== null) {
-      recipeCacheStore.upsertToCache(insertedRecipe)
+      upsertToCache(insertedRecipe)
     }
     return insertedRecipe
   } catch (error) {
@@ -97,7 +126,7 @@ async function updateRecipe(
       newRecipe,
     )
     if (updatedRecipe !== null) {
-      recipeCacheStore.upsertToCache(updatedRecipe)
+      upsertToCache(updatedRecipe)
     }
     return updatedRecipe
   } catch (error) {
@@ -109,7 +138,7 @@ async function updateRecipe(
 async function deleteRecipe(recipeId: Recipe['id']): Promise<void> {
   try {
     await supabaseGateway.deleteRecipe(recipeId)
-    recipeCacheStore.removeFromCache({ by: 'id', value: recipeId })
+    removeFromCache({ by: 'id', value: recipeId })
   } catch (error) {
     logging.error('Recipe error:', error)
   }
