+++ src/modules/measure/infrastructure/tests/measureInfrastructureExports.test.ts
@@ -3,7 +3,6 @@ import { describe, expect, it } from 'vitest'
 import * as measureRepositoryModule from '~/modules/measure/infrastructure/measureRepository'
 import * as realtimeModule from '~/modules/measure/infrastructure/supabase/realtime'
 import * as gatewayModule from '~/modules/measure/infrastructure/supabase/supabaseBodyMeasureGateway'
-import * as mapperModule from '~/modules/measure/infrastructure/supabase/supabaseMeasureMapper'
 
 describe('measure infrastructure module exports', () => {
   it('exports only the measure repository factory', () => {
@@ -15,12 +14,6 @@ describe('measure infrastructure module exports', () => {
   it('exports only the Supabase gateway factory', () => {
     expect(Object.keys(gatewayModule)).toEqual([
       'createSupabaseBodyMeasureGateway',
-    ])
-  })
-
-  it('exports only the Supabase mapper factory', () => {
-    expect(Object.keys(mapperModule)).toEqual([
-      'createSupabaseBodyMeasureMapper',
     ])
   })
