--- src/modules/weight/infrastructure/weight/supabase/constants.ts
@@ -1,1 +1,0 @@
-export const SUPABASE_TABLE_WEIGHTS = 'weights'
