+++ src/modules/weight/infrastructure/weight/supabase/supabaseWeightGateway.ts
@@ -2,11 +2,40 @@ import { type User } from '~/modules/user/domain/user'
 import {
   type NewWeight,
   type Weight,
+  weightSchema,
 } from '~/modules/weight/domain/weight/weight'
 import { type WeightGateway } from '~/modules/weight/domain/weight/weightGateway'
-import { SUPABASE_TABLE_WEIGHTS } from '~/modules/weight/infrastructure/weight/supabase/constants'
-import { supabaseWeightMapper } from '~/modules/weight/infrastructure/weight/supabase/supabaseWeightMapper'
+import { type Database } from '~/shared/supabase/database.types'
 import { supabase } from '~/shared/supabase/supabase'
+import { parseWithStack } from '~/shared/utils/parseWithStack'
+
+const SUPABASE_TABLE_WEIGHTS = 'weights'
+
+type WeightDTO = Database['public']['Tables']['weights']['Row']
+type InsertWeightDTO = Database['public']['Tables']['weights']['Insert']
+type UpdateWeightDTO = Database['public']['Tables']['weights']['Update']
+
+function toDomain(dto: WeightDTO): Weight {
+  return parseWithStack(weightSchema, {
+    ...dto,
+  })
+}
+
+function toInsertDTO(weight: NewWeight): InsertWeightDTO {
+  return {
+    user_id: weight.user_id,
+    weight: weight.weight,
+    target_timestamp: weight.target_timestamp.toISOString(),
+  }
+}
+
+function toUpdateDTO(weight: Weight): UpdateWeightDTO {
+  return {
+    user_id: weight.user_id,
+    weight: weight.weight,
+    target_timestamp: weight.target_timestamp.toISOString(),
+  }
+}
 
 export function createSupabaseWeightGateway(): WeightGateway {
   return {
@@ -28,11 +57,11 @@ async function fetchUserWeights(userId: User['uuid']) {
     throw error
   }
 
-  return weights.map(supabaseWeightMapper.toDomain)
+  return weights.map(toDomain)
 }
 
 async function insertWeight(newWeight: NewWeight) {
-  const weightDTO = supabaseWeightMapper.toInsertDTO(newWeight)
+  const weightDTO = toInsertDTO(newWeight)
   const { data: weight, error } = await supabase
     .from(SUPABASE_TABLE_WEIGHTS)
     .insert(weightDTO)
@@ -43,11 +72,11 @@ async function insertWeight(newWeight: NewWeight) {
     throw error
   }
 
-  return supabaseWeightMapper.toDomain(weight)
+  return toDomain(weight)
 }
 
 async function updateWeight(weightId: Weight['id'], weightUpdate: Weight) {
-  const weightDTO = supabaseWeightMapper.toUpdateDTO(weightUpdate)
+  const weightDTO = toUpdateDTO(weightUpdate)
   const { data: weight, error } = await supabase
     .from(SUPABASE_TABLE_WEIGHTS)
     .update(weightDTO)
@@ -59,7 +88,7 @@ async function updateWeight(weightId: Weight['id'], weightUpdate: Weight) {
     throw error
   }
 
-  return supabaseWeightMapper.toDomain(weight)
+  return toDomain(weight)
 }
 
 async function deleteWeight(id: Weight['id']) {
