+++ src/shared/utils/convertApi2Food.ts
@@ -1,5 +1,5 @@
+import { type ApiFood } from '~/modules/diet/food/domain/apiFood'
 import { createNewFood, type NewFood } from '~/modules/diet/food/domain/food'
-import { type ApiFood } from '~/modules/diet/food/infrastructure/api/domain/apiFoodSchema'
 import { createMacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
 
 /**
